fix(user, DB): 사용자가 자기 자신을 팔로우하지 못하도록 변경

+++ src/main/java/com/ssafy/user/model/service/UserServiceImpl.java
@@ -54,11 +54,17 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public boolean follow(String userId, String targetId) {
+        if (userId.equals(targetId)) {
+            return false;
+        }
         return repo.addFollowing(userId, targetId);
     }
 
     @Override
     public boolean unfollow(String userId, String targetId) {
+        if (userId.equals(targetId)) {
+            return false;
+        }
         return repo.removeFollowing(userId, targetId);
     }
